refactor(tag-pages): extract mergeLocaleDocs to shared utility
Move the duplicated locale-first + base-fallback merge logic out of
both [locale] tag pages into pages/lib/locale-merge.ts. Both pages now
import from the shared module, eliminating the DRY violation flagged
during code review.
