+++ pages/[locale]/docs/tags/index.tsx
@@ -8,22 +8,18 @@
 // map at render time using a locale-doc + base-doc fallback strategy that
 // mirrors src/components/tag-nav.astro's non-default-locale branch.
 //
-// Fallback strategy (locale first, base as fill):
-//   1. Collect docs from `docs-${locale}` collection.
-//   2. Collect base "docs" collection.
-//   3. Merge: locale docs take priority; base docs fill in missing slugs.
+// Fallback strategy (locale first, base as fill): see pages/lib/locale-merge.ts
 //
 // paths() contract (zfb ADR-004 — synchronous):
 //   params: { locale: string }
 //   props:  (none — tag map computed at render time)
 
-import { getCollection } from "zfb/content";
+import { mergeLocaleDocs } from "../../lib/locale-merge";
 import { collectTags } from "@/utils/tags";
 import { toRouteSlug } from "@/utils/slug";
-import { t, defaultLocale } from "@/config/i18n";
+import { t } from "@/config/i18n";
 import { withBase } from "@/utils/base";
 import { settings } from "@/config/settings";
-import type { DocsEntry } from "@/types/docs-entry";
 import { DocLayoutWithDefaults } from "@zudo-doc/zudo-doc-v2/doclayout";
 import { Breadcrumb } from "@zudo-doc/zudo-doc-v2/breadcrumb";
 import type { BreadcrumbItem } from "@zudo-doc/zudo-doc-v2/breadcrumb";
@@ -38,34 +34,6 @@ export function paths(): Array<{ params: { locale: string } }> {
   return Object.keys(settings.locales).map((locale) => ({
     params: { locale },
   }));
-}
-
-/**
- * Merge locale docs with base-locale fallbacks. Locale docs take priority;
- * base docs fill in slugs not covered by the locale collection.
- * Mirrors the non-default-locale branch of tag-nav.astro.
- */
-function mergeLocaleDocs(locale: string): DocsEntry[] {
-  const localeDocs = getCollection(`docs-${locale}`) as unknown as DocsEntry[];
-  const baseDocs = getCollection("docs") as unknown as DocsEntry[];
-
-  const filteredLocale = localeDocs.filter(
-    (d) => !d.data.draft && !d.data.unlisted,
-  );
-  const filteredBase = baseDocs.filter(
-    (d) => !d.data.draft && !d.data.unlisted,
-  );
-
-  const localeSlugSet = new Set(
-    filteredLocale.map((d) => d.data.slug ?? toRouteSlug(d.id)),
-  );
-
-  return [
-    ...filteredLocale,
-    ...filteredBase.filter(
-      (d) => !localeSlugSet.has(d.data.slug ?? toRouteSlug(d.id)),
-    ),
-  ];
 }
 
 interface PageProps {
